# src/main/java/dev/bolohonov/tms/server/dto/UserDto.java
package dev.bolohonov.tms.server.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
@Schema(description = "Сущность пользователя")
public class UserDto {
    /**
     * Имя или логин (email) пользователя
     */
    @Schema(description = "Уникальный логин пользователя (email)")
    private String name;
}